added ack check when receivng data, switched to oneshot125 pwm to control motors

+++ components/motor_control/include/motor_control.h
@@ -1,3 +1,3 @@
 void motor_control_setup();
 void calibrate_esc();
-void set_motor_vals(uint16_t fl, uint16_t fr, uint16_t bl, uint16_t br);
+void set_motor_vals(float fl, float fr, float bl, float br);
